+++ api/CESMII.Marketplace.Data/Entities/RequestInfo.cs
@@ -30,7 +30,7 @@
         /// <summary>
         /// Simple high level information related to an SM profile from CloudLib
         /// </summary>
-        public uint? SmProfileId { get; set; }
+        public long? SmProfileId { get; set; }
 
         public BsonObjectId RequestTypeId { get; set; } //map like statusid for marketplaceitem
